# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PET - Portable Event Telephone API
  version: 1.0.0
  description: REST API for events, extensions, devices, DECT, callbacks and more.
paths:
...
  /api/v1/callback/requests/:
    get:
      operationId: v1_callback_requests_list
      description: Callbacks I requested or that target my extensions; orga see everything
        for ``?event=``.
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCallbackRequestList'
          description: ''
    post:
      operationId: v1_callback_requests_create
      description: Callbacks I requested or that target my extensions; orga see everything
        for ``?event=``.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackRequestRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CallbackRequestRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CallbackRequestRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallbackRequest'
          description: ''
...
  /api/v1/callback/result/:
    post:
      operationId: v1_callback_result_create
      description: 'PBX result hook: ``{event, kind: callback|wakeup|ringback, id,
        result: answered|failed}``.'
      tags:
      - v1
      security:
      - {}
      responses:
        '200':
          description: No response body
  /api/v1/callback/scheduled-calls/:
    get:
      operationId: v1_callback_scheduled_calls_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedScheduledCallList'
          description: ''
    post:
      operationId: v1_callback_scheduled_calls_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledCallRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ScheduledCallRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ScheduledCallRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledCall'
          description: ''
...
  /api/v1/callback/test-ringback/:
    post:
      operationId: v1_callback_test_ringback_create
      description: '``{event: slug, extension: number, delay?: seconds}`` -> ring
        one of my extensions back.'
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/conferences/rooms/:
    get:
      operationId: v1_conferences_rooms_list
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoomList'
          description: ''
    post:
      operationId: v1_conferences_rooms_create
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RoomRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RoomRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
          description: ''
...
components:
  schemas:
...
    AnnouncementEnum:
      enum:
      - default
      - custom
      type: string
      description: |-
        * `default` - Default wake-up announcement
        * `custom` - Custom announcement
...
    CallbackRequest:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
        kind:
          allOf:
          - $ref: '#/components/schemas/KindEnum'
          default: ccbs
        requester_number:
          type: string
          maxLength: 16
        target_number:
          type: string
          maxLength: 16
        state:
          allOf:
          - $ref: '#/components/schemas/CallbackStateEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        expires_at:
          type: string
          format: date-time
          readOnly: true
        attempts:
          type: integer
          readOnly: true
        last_attempt_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        next_attempt_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        channel_id:
          type: string
          readOnly: true
        note:
          type: string
          readOnly: true
      required:
      - attempts
      - channel_id
      - created_at
      - event
      - expires_at
      - id
      - last_attempt_at
      - next_attempt_at
      - note
      - requester_number
      - state
      - target_number
    CallbackRequestRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        kind:
          allOf:
          - $ref: '#/components/schemas/KindEnum'
          default: ccbs
        requester_number:
          type: string
          minLength: 1
          maxLength: 16
        target_number:
          type: string
          minLength: 1
          maxLength: 16
      required:
      - event
      - requester_number
      - target_number
    CallbackStateEnum:
      enum:
      - pending
      - dialing
      - completed
      - cancelled
      - expired
      - failed
      type: string
      description: |-
        * `pending` - Pending
        * `dialing` - Dialing
        * `completed` - Completed
        * `cancelled` - Cancelled
        * `expired` - Expired
        * `failed` - Failed
...
    KindEnum:
      enum:
      - ccbs
      - ccnr
      type: string
      description: |-
        * `ccbs` - Completion of calls to busy subscriber
        * `ccnr` - Completion of calls on no reply
...
    PaginatedCallbackRequestList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/CallbackRequest'
...
    PaginatedRoomList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/Room'
...
    PaginatedScheduledCallList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/ScheduledCall'
...
    RepeatEnum:
      enum:
      - once
      - daily
      type: string
      description: |-
        * `once` - Once
        * `daily` - Daily
...
    Room:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        number:
          type: string
          maxLength: 16
        name:
          type: string
          maxLength: 80
        owner:
          type: string
          readOnly: true
        max_participants:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        record:
          type: boolean
        music_on_hold:
          type: boolean
          description: Play MoH while alone in the room.
        announce_join:
          type: boolean
          description: Play a tone when someone joins/leaves.
        is_public:
          type: boolean
          title: Listed for everyone
        participants:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - number
      - owner
      - participants
    RoomRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
          writeOnly: true
        number:
          type: string
          minLength: 1
          maxLength: 16
        name:
          type: string
          maxLength: 80
        pin:
          writeOnly: true
          description: 4-8 digits; leave blank for an open room.
          oneOf:
          - type: string
            pattern: ^\d{4,8}$
            maxLength: 8
          - type: string
            maxLength: 0
        max_participants:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        record:
          type: boolean
        music_on_hold:
          type: boolean
          description: Play MoH while alone in the room.
        announce_join:
          type: boolean
          description: Play a tone when someone joins/leaves.
        is_public:
          type: boolean
          title: Listed for everyone
      required:
      - event
      - number
...
    ScheduledCall:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        event:
          type: string
        extension:
          type: string
          format: uuid
        extension_number:
          type: string
          readOnly: true
        owner:
          type: string
          readOnly: true
        scheduled_for:
          type: string
          format: date-time
        repeat:
          $ref: '#/components/schemas/RepeatEnum'
        max_retries:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        retry_interval_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        snooze_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        attempts:
          type: integer
          readOnly: true
        state:
          allOf:
          - $ref: '#/components/schemas/ScheduledCallStateEnum'
          readOnly: true
        announcement:
          $ref: '#/components/schemas/AnnouncementEnum'
        announcement_text:
          type: string
          description: Spoken via TTS.
          maxLength: 300
        announcement_file:
          type: string
          format: uri
          nullable: true
        last_result:
          type: string
          readOnly: true
        channel_id:
          type: string
          readOnly: true
        next_attempt_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - attempts
      - channel_id
      - created_at
      - event
      - extension
      - extension_number
      - id
      - last_result
      - next_attempt_at
      - owner
      - scheduled_for
      - state
      - updated_at
    ScheduledCallRequest:
      type: object
      properties:
        event:
          type: string
          minLength: 1
        extension:
          type: string
          format: uuid
        scheduled_for:
          type: string
          format: date-time
        repeat:
          $ref: '#/components/schemas/RepeatEnum'
        max_retries:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        retry_interval_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        snooze_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        announcement:
          $ref: '#/components/schemas/AnnouncementEnum'
        announcement_text:
          type: string
          description: Spoken via TTS.
          maxLength: 300
        announcement_file:
          type: string
          format: binary
          nullable: true
      required:
      - event
      - extension
      - scheduled_for
    ScheduledCallStateEnum:
      enum:
      - scheduled
      - dialing
      - answered
      - failed
      - cancelled
      type: string
      description: |-
        * `scheduled` - Scheduled
        * `dialing` - Dialing
        * `answered` - Answered
        * `failed` - Failed
        * `cancelled` - Cancelled
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid